Music Recommendation System
We have created a Music Recommendation System using a Spotify dataset which is publicly available on Kaggle. Interestingly, Spotify has a web API that developers can use to retrieve audio features and metadata about songs such as the song’s popularity, tempo, loudness, key, and the year in which it was released. We used this data to build music recommendation systems that recommend songs to users based on both the audio features and the metadata of the songs that they have listened to. We have taken the code reference from https://towardsdatascience.com/how-to-build-an-amazing-music-recommendation-system-4cce2719a572.